+++ src/main/java/efub/cpbr/crumble/global/config/WebSecurityConfig.java
@@ -1,6 +1,5 @@
 package efub.cpbr.crumble.global.config;
 
-//import efub.cpbr.crumble.auth.service.UserDetailsServiceImpl; // 사용자 정보 로드 서비스
 import efub.cpbr.crumble.jwt.JwtAuthenticationFilter; // JWT 필터 (지금은 주석 처리)
 import efub.cpbr.crumble.jwt.JwtTokenProvider; // JWT 토큰 생성/검증 도구
 import lombok.RequiredArgsConstructor;
@@ -15,13 +14,13 @@ import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.security.web.SecurityFilterChain;
 import org.springframework.security.web.authentication.UsernamePasswordAuthenticationFilter;
+import org.springframework.web.cors.CorsConfigurationSource;
 
 @Configuration
 @EnableWebSecurity
 @RequiredArgsConstructor
 public class WebSecurityConfig {
 
-    //private final UserDetailsServiceImpl userDetailsService;
     private final JwtTokenProvider jwtTokenProvider;
 
     @Bean
@@ -35,7 +34,7 @@ public class WebSecurityConfig {
     }
 
     @Bean
-    public SecurityFilterChain filterChain(HttpSecurity http) throws Exception {
+    public SecurityFilterChain filterChain(HttpSecurity http, CorsConfigurationSource corsConfigurationSource) throws Exception {
         http
                 .csrf(csrf -> csrf.disable()) // CSRF 보호 비활성화
                 .sessionManagement(session -> session.sessionCreationPolicy(SessionCreationPolicy.STATELESS)) // 세션 사용 안함
@@ -52,8 +51,8 @@ public class WebSecurityConfig {
                         ).permitAll() // 위에 명시된 경로들은 인증 없이 접근 허용
                         .anyRequest().authenticated() // 그 외 모든 요청은 인증 필요
                 )
-
-        .addFilterBefore(new JwtAuthenticationFilter(jwtTokenProvider), UsernamePasswordAuthenticationFilter.class);
+                .cors(cors -> cors.configurationSource(corsConfigurationSource))
+                .addFilterBefore(new JwtAuthenticationFilter(jwtTokenProvider), UsernamePasswordAuthenticationFilter.class);
         return http.build();
     }
 }
